requirements.txt
streamlit
pillow
hachoir